+++ nl.ou.refactoring.advice/src/test/java/nl/ou/refactoring/advice/io/json/JsonSamplesLoader.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public final class JsonSamplesLoader {
 	private JsonSamplesLoader() { }
 
-	protected static String loadJson(String resourceName) {
+	public static String loadJson(String resourceName) {
 		final var jsonResourceStream =
 				JsonSamplesLoader
 					.class