+++ ReleaseNotes.md
@@ -34,3 +34,5 @@
 - Expanded the Refactoring Advice Graph model with more fine-grained code nodes such as blocks inside an operation and statements inside a block.
 - Improved the PlantUML Class Diagram feature: now also supports inner classes.
 - Added support for risk Changed Nested Relationship.
+- Added test case for the "Move Method" refactoring, including JSON and required fixes in the model and flowchart, UML generation.
+- Added JSON Schema generation for serialisation and deserialisation of representations of Refactoring Advice Graphs in JSON.